+++ packages/plugin-sdk/src/errors/connection-closed.error.ts
@@ -1,0 +1,8 @@
+import { NusaClientError } from "./nusa-client.error.js";
+
+export class ConnectionClosedError extends NusaClientError {
+  constructor(message = "WebSocket connection closed") {
+    super("CONNECTION_CLOSED", message);
+    this.name = "ConnectionClosedError";
+  }
+}