+++ packages/plugin-sdk/src/errors/nusa-client.error.ts
@@ -1,0 +1,9 @@
+export class NusaClientError extends Error {
+  readonly code: string;
+
+  constructor(code: string, message: string) {
+    super(message);
+    this.name = "NusaClientError";
+    this.code = code;
+  }
+}